Add support for XDG Base Directory Specification
The Base Directory Specification defines a hierarchy of directories in the
user's home where to place config files etc. The module *File::BaseDir*
provides methods to follow this specification. The default place for the
config file stays *~/.listadmin.ini*. If this file doesn't exist, the XDG
path *~/.config/listadmin.ini* is used.